the simplest possible bytecode interpreter
there are only ~25 LoC here, but it's a whole VM. Unlike the c code presented by Kazanov, there is no instruction pointer because the entire program is loaded into the VM. In any practical VM we'd probably want to pass this by reference. Also unlike Kazanov's, there is less validation needed. The result enum becomes Option. Memory starts zero'd in rust. The compiler ensures there are no bad OpCodes. That last one is another case that may need to be changed in a practical VM.