embrace and extend - second simplest bci
building off the last interpreter we clone it into a new library, the worlds second simplest bytecode interpreter (wssbi). This extends the first by adding arguments for opcodes. in kazanovs example this is done by adding opcodes that can increment the pointer and read ahead one space. Sternberg's rust articles use this to illustrate the enhanced power of rust enums here. Since enum is a sum type it can carry information! This necessarily diverges from Kazanov's internal representation which relies on C enums actually being integers under the hood. There's a bit of undefined behavior here in the C++03 standard, where the exact integral type is left to the compiler. This lets you put other integrals in the programs IR and then coerce them as needed.

In this commit I've set up the new instructions as noops and left the implementation undecided. I'm tempted to jump right to implementing a stack here.