leverage rust sum types in opcodes
after thinking about it for a few minutes I decided I liked sternbergs approach and had no practice with enums as sum types so I should try it out. This representation of the program would be produced by some other parser ahead of time, anyway, and like the compromises made in wsbi earlier we are trading off what a practical interpreter might have for simplicity.

Implementing a stack is a right answer here, and we'll get to that next, but this is nice because it leans harder into the rust language features to give us a more generalized instruction set.