add a stack and arguments
this builds off the work of wssbi. we move from a single register to a stack which can contain several values, and rewrite our opcodes to assume the stack. we also add an enum for possible program errors and change the structure of the main loop to allow for propogating these program errors. defining how to handle each of the opcodes is tedious so I've borrowed sternberg's macro for removing the boilerplate.

now any program passed in will halt as soon as an error occurs, and otherwise will run until a Done opcode or it's out of instructions. or at least, it should. One side effect of rewriting the opcodes was losing Inc, Dec, and changing the IR representation of Add, Sub so all the tests need to be changed. The VM is also now returning a result instead of an option type. We've got a single test now demonstrating this.

We're also not handling some pretty obvious error states or testing any of our new error handling, so this machine isn't quite done.