add tests, removed popres
this adds happy-path tests evaluating each of the opcodes, including a few complex chains at the end.

Kazinov's instruction set here preserves a single register and has a single opcode for popping the top value off the stack and putting that top value into the register. It's kind of goofy. There are no instructions for reading the register.

Like Sternberg I'm omiting it here. The next generation of this interpreter will implement registers, and I like the forth-like aesthetic of this toy.

Next I'll handle some new Program Errors at runtime, at which point this is probably done and we can move on to the next one.