switch to match style
if let is syntactic sugar for match and best applied when we only want to act on one of the two possible cases. Since we handle both here, I find the match style to be easier to read in this instance, although it's nested deeper now.

In python or any other c-style language, though, I would try to return early with the error code in more places. I'm not sure if that's idiomatic in rust, though.