implement opcodes
expecting code to not compile and commiting it anyway, with an uncritical eye towards the failing compiler output, is asking for trouble.

This implements most of the opcodes necessary to get our register machine to work. It also fixes several misplaced closing brackets and parens, and two instances of variable name confusion. hooray for static checks.

Even though this compiles and the single test passes, though, not all the opcodes are excercies. Like the stack machine, there are hidden bugs in here that will be addressed in the next fix as I add more tests.