all the register tests
this adds a more exhaustive suite of tests of the register machine. The tests explicitly reveal two issues with division. First, that we were not handling division by zero (the Error was there but unused. It came over from the initial copy of vm 03's source). Second that Rust's underlying 'truncation toward zero' property leaks through. This isn't necessarily a problem as our VM is not well defined. More of a curiousty that the tests spell out clearly.

Sternberg points out that this VM is not doing sufficient error checking for us. Uninitialized memory is not tracked (indeed, this patch adds tests to demonstrate memory is initialized at 0), nor is out of bounds access handled well (it panics).

In future work, the macro pattern from vm 03 could add bounds checking to these opcodes. Also a lot of these tests could more concisely replaced with quick check.