finish instructions and add bounds checking
at this point we have, more or less, the worst regex machine that can actually be called such. It doesn't support non-ascii very well, as char's in rust are only a single unicode point and multiple chars may be necessary for single characters in unicode. There may even be parts of ascii it doesn't support well. We won't be adding this since it is non-trivial and would easily eclipse the whole excercise. This also will not handle the closure operator (*).

Next up, tests will help define the limits of the system.