add readme with overview
This was a fun exploration, and looking back on it the whole thing certainly felt larger than it appears now that it's done. I may come back around to write the lc-3 VM part eventually, but realistically I'm going to take a break from VMs for a bit.

If you're reading this, you may find some insight in the commit log. I tend to leave verbose commit messages so that it's easier to pick up where I left off. I summarized these notes in the readme descriptions of each VM but that does not capture the same motion as the intra-development notes in the log.